Serve retriever queries in bfloat16, not float16
ielabgroup/ITER-Qwen3-Embedding-0.6B's model card requires bfloat16:
float16 produces measurably different ranking results.

+++ run_eval.py
@@ -97,7 +97,7 @@ def main():
     if args.searcher != "bm25":
         cmd += ["--model-name", retriever,
                 "--dataset-name", bench.corpus,
-                "--pooling", "eos", "--normalize", "--torch-dtype", "float16",
+                "--pooling", "eos", "--normalize", "--torch-dtype", "bfloat16",
                 "--task-prefix", setting.prefix,
                 "--max-length", str(setting.max_length)]
